separate components on folders
